Title: 몫 출력하기, Time: 49ms, Memory: 16MB, Status: Passed - CodeTree

+++ 250322/몫 출력하기/print-share.py
@@ -3,7 +3,7 @@ prod = 0
 while True :
     a = int(input())
     if a % 2 == 0 :
-        print(a / 2)
+        print(int(a / 2))
         prod += 1
         if prod == 3 :
             break
